services/api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
python-dateutil==2.9.0.post0
pydantic==2.7.4
numpy==1.26.4
pandas==2.2.2

python-multipart==0.0.9
openai>=1.0.0
python-dotenv>=1.0.1
